+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Change categories:
 1. base: Stabilise pldm_msg_hdr_correlate_response()
 2. transport: af-mctp: Stabilise pldm_transport_af_mctp_bind()
 3. libpldm: Fix header use
+4. libpldm: More fixes for header use
 
 ### Fixed